requirements.txt
fastapi==0.108.0
line_bot_sdk==3.5.1
matplotlib==3.8.2
mplfinance==0.12.10b0
opencv_python==4.8.1.78
pandas==2.1.4
Pillow==10.1.0
python-dotenv==1.0.0
ultralytics==8.0.227
yfinance==0.2.33